core/sip/trans_layer.cpp: eliminated one compiler warning
- Don't know if dereferencing warnings could also be eliminated.

+++ core/sip/trans_layer.cpp
@@ -847,7 +847,7 @@ int _trans_layer::send_request(sip_msg* msg, trans_ticket* tt,
 	    return res;
 	}
 
-	DBG("dest_list.size() = %lu\n",dest_list.size());
+	DBG("dest_list.size() = %lu\n", (long unsigned int)dest_list.size());
 
 	if(dest_list.size() == 1) {
 	    next_hop = dest_list.front().host;
